requirements.txt
git+https://github.com/openai/whisper.git
torch
torchaudio
pyannote.audio
ffmpeg-python
sox
transformers
accelerate
requests
python-dotenv
